feat: #30 Changing function structure and adding OCRProcessor class
